Update Mobs: MovementControl and sprite animations Add Thai medicine description

+++ Data/Models/CharacterData.cs
@@ -5,8 +5,8 @@
         public int CharId { get; set; }
         public string Name { get; set; }
         public int Category { get; set; }
-        public int ATK { get; set; }
-        public int HP { get; set; }
+        public double ATK { get; set; }
+        public double HP { get; set; }
         public double DEF_Percent { get; set; }
         public double Crit_Percent { get; set; }
         public double CritDMG_Percent { get; set; }
